removed imsize=14400 from generic scripts

+++ IRAS04302_spectral_line_SBLB.py
@@ -145,6 +145,7 @@ with open(prefix+'.pickle', 'wb') as handle:
 ###############################################################
 
 for i in data_params.keys():
+   if 'LB' in i:
       os.system('rm -rf '+data_params[i]['vis_contsub']+'.tgz')
       os.system('tar czf '+data_params[i]['vis_contsub']+'.tgz '+data_params[i]['vis_contsub'])
